requirements.txt
boto3==1.6.18
botocore==1.9.18
Django==1.9.5
djangorestframework==3.3.3
django-rest-swagger==0.3.6
django-cors-headers==2.1.0
whitenoise==3.3.0
requests==2.19.1
